CI (manylinux): Added job to build manylinux2014 wheels
Until now, the workflow only had jobs to build MacOS wheels and source
distributions. However, the main format of distributing Python packages
is through Python wheels. And to ensure that the package reaches a wider
audience amongst linux users, we will make use of manylinux wheels.

With this commit, we add a new job to build manylinux2014_x86_64 wheels.
We make use of a custom pyGnuastro docker which already has Gnuastro and
its dependencies built. The job only runs the build-wheels script from
within the container and returns the wheels produced as artifacts.

Also made few changes to the sdist and MacOS jobs to handle errors:
- Added some checks in sdist to verify gnuastro installation.
- Added ldconfig in sdist.
- Reduced the python versions in MacOS wheels, since some of the newer
  versions were throwing errors.
- Changed install instructions for MacPorts.