CI (MacOS): Switched from MacPorts to conda in MacOS job
Until now, the MacOS wheel builder job tried using the MacPorts package
manager for installing gnuastro and its dependencies. However, this posed
multiple errors. Moreover MacOS runners provided by GitHub actions come
with anaconda installed.

With this commit, we use `conda forge` instead of MacPorts for installing
gnuastro. Also supplied the necessary flags to `setup.py build_ext` call
to link the gnuastro libraries and include header files in the conda
directories.

We also switch from `gnuastro` to `libgnuastro-dev` as the package to be
installed for the source dist build. Since the sdist depends on the
gnuastro library only, installing `libgnuastro-dev` makes it easier to
link it to the library then the whole `gnuastor` package.

We also switch the Python version used to build the sdist from `3.6` to
`3.8`.