Test, Setup : Tests for fits module, conditional build on MacOS
Until now, we had no tests for checking the FITS image read and write
functions.

With this commit, we add a new `test_fits.py` file which has classes to
test these functions on an example FITS image(`psf.fits`).

NOTE: The test is passed upon generation of an output. The value of the
      output is not checked.

Until now, we used `conda forge` to install gnuastro on MacOS runners in
the workflow. However, this restricts us to latest stable release version
i.e 0.18. Since this version doesnt contain the Python interface(python.h)
functions, we encounter errors if we try to build the modules which depend
on them(fits module).

With this commit, we add a script in `setup.py` to not include the modules
which depend on the Python interface functions while building for MacOS.