Changed error handling in FITS module
Until now, error handling in the FITS module was done using the Gnulib
error module. However, this module is not available on all platforms.

With this commit, we remove the dependency on this module thus making
the package more portable. We also add some Python error handling to assist
this