Setup (dev): Changed version to dev for TestPyPI
Until now, the version of pyGnuastro was in the normal
python version format (eg 0.0.1). However, for uploads
on TestPyPI, we need to have a separate versioning.

With this commit, we add a `-dev` to the versions
(eg 0.0.1-dev0). And the development versions for
a new release will have their own versioning indicated
by the number after `-dev`.