Numpy: Fixed import_array errors
Until now, import_array() was called ONLY in the module initialization
function of the FITS module. However, there was no call to import_array()
in the `utils.c` file, which resulted in "defined but not used" warnings
from numpy.

With this commit, we add a PY_ARRAY_UNIQUE_SYMBOL macro to both
files. This has to be included so that the NumPy C-API can be included
in multiple files for a single-extension module. The import_array()
function included in the module initialization, is declare static
if this macro is not defined. This means that the NumPy C-API doesn't
extend to other files (like utils.c) without this macro.

Also, defined NO_IMPORT_ARRAY in the `utils.c` file to not cause the
warnings.