Documentation: Added documentation source for Website
Until now, we had no documentation for the pyGnuastro package outside of
the README.

With this commit, we add all the necessary source required for a website.
The documentation will be built using Sphinx and Autodoc. All the files
related to the documentation will be stored in the `docs/` folder. Also
added `sphinx` to dev-requirements.

Also, added a 'gh-pages' branch to host the documentation.

With this commit, we also add a condition to only trigger workflow runs when
a push is made on the main branch and to not trigger a run incase of changes
only in `docs/`